src/requirements.txt
watchdog
python-dotenv
mysql-connector-python